backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pillow
google-generativeai
torch
torchvision
transformers
pytorch-lightning
opencv-python
grad-cam
scikit-learn
numpy
